Adding initial support for OAuth-based authentication.
OAuth authentication requires the Python oauth module.  A unit test will be
added soon so that this additional authentication path is also exercised.

In addition to the OAuth-specific code, this change also restructures the HTTP
path a bit more.  We now pass request parameters separately as dictionaries to
allow them to be used as part of the OAuth request building process.

Also, httplib debug output can be enabled using the new 'debug' parameter to
the Brightkite object.